+++ lib/src/code_generators/constants.dart
@@ -12,7 +12,7 @@ const kBasicTypes = [
   'float',
   'bool',
   'boolean',
-  'number',
+  'num',
 ];
 
 const kChopperService = 'ChopperService';